src/requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
fastapi
tflite
tensorflow
tflite-runtime
tensorflow-hub
